requirements.txt
anyio==3.6.1
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ek9yztvu3/croot/certifi_1665076692562/work/certifi
click==8.1.3
dnspython==2.2.1
fastapi==0.85.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.0.0
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.3
